clutter: Move ClutterActorPrivate to clutter-muffin.h
This will be needed to override queue_relayout on MetaWindowActor without hurting performance.
